# Allocate CKF links in a flat allocation