--- a/device/common/include/traccc/finding/device/impl/propagate_to_next_surface.ipp
+++ b/device/common/include/traccc/finding/device/impl/propagate_to_next_surface.ipp
@@ -38,7 +38,8 @@ TRACCC_DEVICE inline void propagate_to_next_surface(
     vecmem::device_vector<const candidate_link> links(payload.links_view);
 
     // Seed id
-    unsigned int orig_param_id = links.at(param_id).seed_idx;
+    unsigned int orig_param_id =
+        links.at(payload.prev_links_idx + param_id).seed_idx;
 
     // Count the number of tracks per seed
     vecmem::device_atomic_ref<unsigned int> num_tracks_per_seed(
@@ -54,11 +55,12 @@ TRACCC_DEVICE inline void propagate_to_next_surface(
     }
 
     // tips
-    vecmem::device_vector<candidate_tip> tips(payload.tips_view);
+    vecmem::device_vector<unsigned int> tips(payload.tips_view);
 
-    if (links.at(param_id).n_skipped > cfg.max_num_skipping_per_cand) {
+    if (links.at(payload.prev_links_idx + param_id).n_skipped >
+        cfg.max_num_skipping_per_cand) {
         params_liveness[param_id] = 0u;
-        tips.push_back({payload.step, param_id});
+        tips.push_back(payload.prev_links_idx + param_id);
         return;
     }
 
@@ -109,7 +111,7 @@ TRACCC_DEVICE inline void propagate_to_next_surface(
         params[param_id] = propagation._stepping.bound_params();
 
         if (payload.step == cfg.max_track_candidates_per_track - 1) {
-            tips.push_back({payload.step, param_id});
+            tips.push_back(payload.prev_links_idx + param_id);
             params_liveness[param_id] = 0u;
         } else {
             params_liveness[param_id] = 1u;
@@ -118,7 +120,7 @@ TRACCC_DEVICE inline void propagate_to_next_surface(
         params_liveness[param_id] = 0u;
 
         if (payload.step >= cfg.min_track_candidates_per_track - 1) {
-            tips.push_back({payload.step, param_id});
+            tips.push_back(payload.prev_links_idx + param_id);
         }
     }
 }