--- a/device/common/include/traccc/finding/device/propagate_to_next_surface.hpp
+++ b/device/common/include/traccc/finding/device/propagate_to_next_surface.hpp
@@ -14,7 +14,6 @@
 #include "traccc/definitions/qualifiers.hpp"
 #include "traccc/edm/track_parameters.hpp"
 #include "traccc/finding/candidate_link.hpp"
-#include "traccc/finding/candidate_tip.hpp"
 #include "traccc/finding/finding_config.hpp"
 
 // VecMem include(s).
@@ -57,6 +56,11 @@ struct propagate_to_next_surface_payload {
     vecmem::data::vector_view<const candidate_link> links_view;
 
     /**
+     * @brief Index in the link vector at which the current step starts
+     */
+    const unsigned int prev_links_idx;
+
+    /**
      * @brief Current CKF step number
      */
     unsigned int step;
@@ -69,7 +73,7 @@ struct propagate_to_next_surface_payload {
     /**
      * @brief View object to the vector of tips
      */
-    vecmem::data::vector_view<candidate_tip> tips_view;
+    vecmem::data::vector_view<unsigned int> tips_view;
 
     /**
      * @brief View object to the vector of the number of tracks per initial