… core/include/traccc/finding/candidate_tip.hpp
… device/cuda/src/finding/finding_algorithm.cu
… device/sycl/src/finding/find_tracks.hpp