This commit migrates the link allocation in the combinatorial Kálmán filter from using a jagged buffer to using a contiguous, flat buffer. This has several advantages:

1. It generally reduces the complexity of the code.
2. It reduces the number of allocations and copies in the code.
3. It removes the usage of awkward jagged vectors.
4. It increases memory efficiency.
5. It provides kernels with the ability to view links from all steps, not just the current and previous one.