@stephenswat 2025-03-25
> Apart from our SoA discussion, which I still don't feel super strongly about (so the PR's code may just as well go ahead), do you have some results for what it does to the performance of the code? Apart from testing that it would still find the exact same tracks that it did so far... 🤔

Let me see if pcadp04 is free and I'll make some!