@stephenswat 2025-03-27
### Performance changes

![924](https://github.com/user-attachments/assets/3fb0fda6-8dca-4ba9-95af-65181e39bbec)

Note that this doesn't take into account performance changes outside of kernels. Throughput example results:

### Before

```
$ ./build/bin/traccc_throughput_st_cuda --input-directory=/data/Acts/odd-simulations-20240506/geant4_ttbar_mu200 --use-detray-detector=on --input-events=10 --cold-run-events 10 --processed-events 100 --use-acts-geom-source=on --deterministic
Reconstructed track parameters: 8846209
Time totals:
                  File reading  5142 ms
            Warm-up processing  8292 ms
              Event processing  50901 ms
Throughput:
            Warm-up processing  829.291 ms/event, 1.20585 events/s
              Event processing  509.015 ms/event, 1.96458 events/s
```

### After

```
$ ./build/bin/traccc_throughput_st_cuda --input-directory=/data/Acts/odd-simulations-20240506/geant4_ttbar_mu200 --use-detray-detector=on --input-events=10 --cold-run-events 10 --processed-events 100 --use-acts-geom-source=on --deterministic
Reconstructed track parameters: 8846324
Time totals:
                  File reading  5456 ms
            Warm-up processing  181880 ms
              Event processing  50846 ms
Throughput:
            Warm-up processing  18188.1 ms/event, 0.0549811 events/s
              Event processing  508.465 ms/event, 1.9667 events/s
```